# api.yml
openapi: 3.0.3
info:
  title: api
  description: api
  version: 0.0.1
servers:
  - url: http://0.0.0.0:3000
    description: api
...
paths:
...
  /receptions:
    get:
      tags:
        - api
      summary: Get receptions
      description: Get receptions
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
    post:
      tags:
        - api
      summary: Add/Update reception
      description: Add/Update reception
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                doctorId:
                  type: integer
                patientId:
                  type: integer
                date:
                  type: string
                  format: date
                startTime:
                  type: string
                endTime:
                  type: string
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
  /reception/{receptionId}:
    get:
      tags:
        - api
      summary: Get reception
      description: Get reception
      parameters:
        - name: receptionId
          in: path
          required: true
          description: reception id
          schema:
            type: integer
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
    delete:
      tags:
        - api
      summary: Remove reception
      description: Remove reception
      parameters:
        - name: receptionId
          description: reception id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
  /receptions/interval:
    post:
      tags:
        - api
      summary: Add/Update receptions interval
      description: Add/Update receptions interval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                doctorId:
                  type: integer
                patientId:
                  type: integer
                date:
                  type: string
                  format: date
                startInterval:
                  type: string
                endInterval:
                  type: string
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reception'
  /reception/take/{receptionId}:
    put:
      tags:
        - api
      summary: Take reception
      description: Take reception
      parameters:
        - name: receptionId
          description: reception id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patientId:
                  type: integer
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/response'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Reception'
components:
  schemas:
    'response':
      type: object
      properties:
        status:
          type: integer
        success:
          type: boolean
        data:
          anyOf:
            - type: object
            - type: array
        message:
          type: string
        code:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: integer
        role_id:
          type: integer
        firstname:
          type: string
        middlename:
          type: string
        lastname:
          type: string
        gender:
          type: string
          enum: [male, female]
        age:
          type: integer
        phone:
          type: string
    Reception:
      type: object
      properties:
        id:
          type: integer
        doctor_id:
          type: integer
        patient_id:
          type: integer
        date:
          type: string
          format: date
        start_time:
          type: string
        end_time:
          type: string